+++ probhub/cli.py
@@ -239,7 +239,7 @@ def _init_typst_templates(root, subtitle):
     typst_dir = typst_root / subtitle
     sources = {
         typst_root / "lib.typ": references / "lib.typ",
-        typst_root / "probhub.svg": package_root / "logo.svg",
+        typst_root / "school-badge.png": package_root / "school-badge.png",
         typst_dir / "main.typ": references / "main.typ",
         typst_dir / "problems.typ": references / "problems.typ",
     }
@@ -323,7 +323,7 @@ def command_init(args):
                 "directory": f"typst-statement/{subtitle}",
                 "creation_timestamp": 0,
                 "cover": {
-                    "logo": "probhub.svg",
+                    "logo": "school-badge.png",
                     "logo_width": "9cm",
                     "logo_space_above": "0em",
                     "logo_space_below": "0em",